added streaming.py and test_streaming.py preliminary setup for annotation of CESR streams
