+++ tests/core/test_streaming.py
@@ -1,0 +1,37 @@
+# -*- coding: utf-8 -*-
+"""
+tests.core.test_streaming module
+
+"""
+
+
+import pytest
+
+
+from keri import kering
+
+from keri.help import helping
+
+from keri.core import (Matter, )
+
+
+from keri.core import streaming
+from keri.core.streaming import (annot, denot, Streamer)
+
+
+def test_annot():
+    """Test annot function  Annotate"""
+
+    """End Test"""
+
+def test_streamer():
+    """Test streamer instance"""
+
+    """End Test"""
+
+if __name__ == "__main__":
+    test_annot()
+    test_streamer()
+
+
+
